requirements-test.txt
pytest>=8.0.0
pytest-homeassistant-custom-component>=0.13.110
pymodbus>=3.8.0
